Refactor ConsensusMessage component by removing unused imports and extracting ModelResponseCard logic into a separate component for improved reusability and maintainability.
